# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login exitoso. Puede requerir 2FA si está habilitado.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/LoginResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Credenciales inválidas o usuario inactivo",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "example": null
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "429": {
            "description": "Demasiados intentos fallidos",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "example": null
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Login de usuario",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Token renovado exitosamente. El refresh token se rota automáticamente en cookie.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/RefreshResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Refresh token inválido, revocado o expirado",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "example": null
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Renovar access token (con rotation)",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "204": {
            "description": "Logout exitoso. Tokens revocados y cookie limpiada.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "example": null
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "example": null
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Logout de usuario",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "description": "Retorna la información completa del usuario actual incluyendo datos de su organización y rutas de navegación basadas en sus roles",
        "operationId": "AuthController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Perfil del usuario con navegación",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UserResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "example": null
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Usuario no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "example": null
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Obtener perfil del usuario autenticado con navegación",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/switch-role": {
      "post": {
        "description": "Permite al usuario cambiar entre sus roles asignados. El cambio persiste en todas las sesiones activas y futuros refreshes.",
        "operationId": "AuthController_switchRole",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwitchRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Rol cambiado exitosamente. Nuevo access token generado.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/SwitchRoleResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "example": null
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "description": "El usuario no tiene asignado el rol solicitado",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "example": null
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cambiar rol activo del usuario",
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "ATR API",
    "description": "API de auditorías, plantillas, frameworks de madurez y gestión de usuarios",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "SuccessResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Indica si la operación fue exitosa",
            "example": true,
            "default": true
          },
          "statusCode": {
            "type": "number",
            "description": "Código de estado HTTP",
            "example": 200,
            "enum": [
              200,
              201,
              204
            ]
          },
          "message": {
            "type": "string",
            "description": "Mensaje descriptivo de la operación. Se genera automáticamente según el método HTTP o se personaliza con @ResponseMessage",
            "example": "Operación exitosa"
          },
          "data": {
            "type": "object",
            "description": "Datos de la operación. Será null si el endpoint no retorna datos.",
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "description": "Timestamp ISO 8601 de cuando se generó la respuesta",
            "example": "2024-01-30T10:30:00.000Z"
          }
        },
        "required": [
          "success",
          "statusCode",
          "message",
          "timestamp"
        ]
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "Access token JWT (válido por 15 minutos). Vacío si requiere 2FA. Usar este token para llamar a /auth/me y obtener los datos del usuario.",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
          },
          "require2FA": {
            "type": "boolean",
            "description": "Indica si se requiere verificación de código 2FA. Si es true, el accessToken estará vacío hasta verificar el código. El frontend debe mostrar la pantalla de verificación 2FA.",
            "example": false
          },
          "twoFactorToken": {
            "type": "string",
            "description": "Token temporal para validar el código 2FA. Solo presente cuando require2FA=true. El cliente debe enviarlo en POST /auth/2fa/verify junto con el código recibido por email.",
            "example": "a1b2c3d4e5f6789012345678901234567890123456789012345678901234abcd"
          }
        },
        "required": [
          "accessToken"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "usernameOrEmail": {
            "type": "string",
            "description": "Email o username del usuario",
            "example": "admin@gmail.com",
            "maxLength": 255
          },
          "password": {
            "type": "string",
            "description": "Contraseña del usuario",
            "example": "Password123!",
            "maxLength": 128
          },
          "rememberMe": {
            "type": "boolean",
            "description": "¿Recuérdame? (opcional)",
            "example": true,
            "default": false
          }
        },
        "required": [
          "usernameOrEmail",
          "password"
        ]
      },
      "RefreshResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "Nuevo access token JWT. El refresh token se rota automáticamente y se envía en cookie HTTP-only.",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": [
          "accessToken"
        ]
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID único del usuario",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "names": {
            "type": "string",
            "description": "Nombres del usuario",
            "example": "Juan Carlos"
          },
          "lastNames": {
            "type": "string",
            "description": "Apellidos del usuario",
            "example": "Pérez García"
          },
          "email": {
            "type": "string",
            "description": "Email del usuario",
            "example": "juan.perez@example.com"
          },
          "username": {
            "type": "string",
            "description": "Nombre de usuario único",
            "example": "juan_perez"
          },
          "ci": {
            "type": "string",
            "description": "Cédula de identidad",
            "example": "1234567-8"
          },
          "phone": {
            "type": "object",
            "description": "Teléfono de contacto",
            "example": "+591 2 234 5678",
            "nullable": true
          },
          "address": {
            "type": "object",
            "description": "Dirección del usuario",
            "example": "Av. 6 de Agosto #123",
            "nullable": true
          },
          "image": {
            "type": "object",
            "description": "URL de la imagen de perfil",
            "example": "https://example.com/images/profile.jpg",
            "nullable": true
          },
          "isActive": {
            "type": "boolean",
            "description": "Estado del usuario",
            "example": true
          },
          "isTemporaryPassword": {
            "type": "boolean",
            "description": "Indica si el usuario debe cambiar su contraseña temporal",
            "example": true
          },
          "firstLoginAt": {
            "type": "object",
            "description": "Fecha y hora del primer login del usuario",
            "example": "2024-01-15T10:30:00.000Z",
            "nullable": true
          },
          "isTwoFactorEnabled": {
            "type": "boolean",
            "description": "Indica si el usuario tiene 2FA habilitado",
            "example": false
          },
          "roles": {
            "type": "array",
            "description": "Roles asignados al usuario",
            "example": [
              "auditor"
            ],
            "items": {
              "type": "string",
              "enum": [
                "admin",
                "gerente",
                "auditor",
                "cliente"
              ]
            }
          },
          "organizationId": {
            "type": "string",
            "description": "ID de la organización a la que pertenece",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "organizationName": {
            "type": "string",
            "description": "Nombre de la organización",
            "example": "Empresa S.A."
          },
          "organizationImage": {
            "type": "object",
            "description": "URL de la imagen de la organización",
            "example": "https://example.com/images/org-logo.jpg",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Fecha de creación del usuario",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Fecha de última actualización",
            "example": "2024-01-15T10:30:00.000Z"
          }
        },
        "required": [
          "id",
          "names",
          "lastNames",
          "email",
          "username",
          "ci",
          "isActive",
          "isTemporaryPassword",
          "isTwoFactorEnabled",
          "roles",
          "organizationId",
          "organizationName",
          "createdAt",
          "updatedAt"
        ]
      },
      "SwitchRoleResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "Nuevo access token con el rol activo cambiado",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "currentRole": {
            "type": "string",
            "description": "Rol activo actual después del cambio",
            "enum": [
              "admin",
              "gerente",
              "auditor",
              "cliente"
            ],
            "example": "auditor"
          },
          "availableRoles": {
            "type": "array",
            "description": "Lista de todos los roles disponibles del usuario",
            "example": [
              "gerente",
              "auditor"
            ],
            "items": {
              "type": "string",
              "enum": [
                "admin",
                "gerente",
                "auditor",
                "cliente"
              ]
            }
          }
        },
        "required": [
          "accessToken",
          "currentRole",
          "availableRoles"
        ]
      },
      "SwitchRoleDto": {
        "type": "object",
        "properties": {
          "newRole": {
            "type": "string",
            "description": "Nuevo rol activo que el usuario quiere usar",
            "enum": [
              "admin",
              "gerente",
              "auditor",
              "cliente"
            ],
            "example": "auditor"
          }
        },
        "required": [
          "newRole"
        ]
      },
...
    }
  }
}